# docs/contracts/openapi/core-api-v1.yaml
openapi: 3.1.0
info:
  title: CogniForge Core API
  description: '# CogniForge API-First Architecture


    منصة تعليمية ذكية مدعومة بالذكاء الاصطناعي مع معمارية API-First كاملة.


    ## المبادئ الأساسية

    - **API-First**: النظام مصمم أولاً كـ API

    - **Platform Independent**: يعمل على أي منصة (Web, Mobile, Desktop)

    - **Language Agnostic**: يمكن استخدامه من أي لغة برمجة

    - **Stateless**: كل طلب مستقل


    ## المصادقة

    يستخدم النظام JWT (JSON Web Tokens) للمصادقة.

    ```

    Authorization: Bearer <token>

    ```


    ## معدلات الطلبات

    - **معدل عام**: 100 طلب/دقيقة

    - **نقاط النهاية الحساسة**: 10 طلبات/دقيقة


    ## Versioning

    - **الإصدار الحالي**: v1

    - **Base URL**: `/api/v1`

    '
  version: 1.0.0
  contact:
    name: CogniForge Support
    email: support@cogniforge.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
- url: https://api.cogniforge.ai/api/v1
  description: Production Server
- url: https://staging-api.cogniforge.ai/api/v1
  description: Staging Server
- url: http://localhost:8000/api/v1
  description: Local Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/security/login
...
paths:
...
  /admin/api/hard-currency/frontier:
    get:
      tags:
      - Admin
      summary: Hard-Currency Frontier
      operationId: adminHardCurrencyFrontier
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Value-chain frontier re-derived from the contact ledger (D-305)
        '503':
          description: A source file is absent from this deployment
      description: Hard-Currency Frontier
  /admin/api/hard-currency/einvoicing-audits:
    post:
      tags:
      - Admin
      summary: E-Invoicing Audit
      operationId: adminHardCurrencyEinvoicingAudit
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
              - file
              - corridor
              properties:
                file:
                  type: string
                  format: binary
                corridor:
                  type: string
                  enum:
                  - fr
                  - be
      responses:
        '200':
          description: Deterministic audit, in memory, nothing stored
        '413':
          description: File too large
        '415':
          description: Not a CSV file
        '422':
          description: Unreadable CSV or unknown corridor
      description: E-Invoicing Audit
  /admin/api/hard-currency/cbam/codes:
    get:
      tags:
      - Admin
      summary: CBAM Pinned Codes
      operationId: adminHardCurrencyCbamCodes
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Pinned CN codes with provenance
      description: CBAM Pinned Codes
  /admin/api/hard-currency/cbam/codes/{cn}:
    get:
      tags:
      - Admin
      summary: CBAM Code Detail
      operationId: adminHardCurrencyCbamCodeDetail
      security:
      - BearerAuth: []
      parameters:
      - name: cn
        in: path
        description: Pinned CN code.
        required: true
        schema:
          type: string
      - name: year
        in: query
        required: false
        schema:
          type: integer
          default: 2026
      responses:
        '200':
          description: Pinned default, crossover threshold and trajectory
        '404':
          description: CN code is not pinned
        '422':
          description: Year outside the pinned horizon
      description: CBAM Code Detail
  /admin/api/hard-currency/cbam/codes/{cn}/decision:
    post:
      tags:
      - Admin
      summary: CBAM Decision
      operationId: adminHardCurrencyCbamDecision
      security:
      - BearerAuth: []
      parameters:
      - name: cn
        in: path
        description: Pinned CN code.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - see_actual
              properties:
                see_actual:
                  type: number
                  exclusiveMinimum: 0
                  maximum: 50
      responses:
        '200':
          description: First year declaring actual data is cheaper, or none within the horizon
        '404':
          description: CN code is not pinned
        '422':
          description: Measured emissions out of range or code not computable
      description: CBAM Decision
  /admin/api/hard-currency/redteam/classes:
    get:
      tags:
      - Admin
      summary: AR/FR Exploit Classes
      operationId: adminHardCurrencyRedteamClasses
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Exploit classes with publication status; probe texts are never returned
      description: AR/FR Exploit Classes
  /admin/api/hard-currency/chamber:
    get:
      tags:
      - Admin
      summary: Decision Chamber
      operationId: adminHardCurrencyChamber
      security:
      - BearerAuth: []
      responses:
        '200':
          description: Highest truthful claim, first missing proof and one human action, each
            sentence classified with its evidence ids (D-306)
        '503':
          description: A source is absent, or a sentence outran its evidence and the chamber
            refused to speak
      description: Decision Chamber
  /admin/api/hard-currency/chamber/cross-examination:
    post:
      tags:
      - Admin
      summary: Decision Chamber Cross-Examination
      operationId: adminHardCurrencyChamberCrossExamination
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - question
              properties:
                question:
                  type: string
                  enum:
                  - ready
                  - build
                  - why_no_money
                  - say_to_buyer
                text:
                  type: string
                  maxLength: 2000
      responses:
        '200':
          description: Classified sentences; for say_to_buyer, the verdict of the D-304 rules
        '422':
          description: A question outside the closed set, or say_to_buyer without text
        '429':
          description: Rate limited
      description: Decision Chamber Cross-Examination
  /admin/api/hard-currency/chamber/outcome-preview:
    post:
      tags:
      - Admin
      summary: Decision Chamber Outcome Preview
      operationId: adminHardCurrencyChamberOutcomePreview
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - date
              - target_ref
              - entity
              - country
              - channel
              - action
              properties:
                date:
                  type: string
                target_ref:
                  type: string
                entity:
                  type: string
                country:
                  type: string
                channel:
                  type: string
                action:
                  type: string
                amount_eur:
                  type: string
                evidence_ref:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: The proposed ledger row accepted or rejected with its reasons, the CSV
            line and the delta; nothing is written (written=false)
        '429':
          description: Rate limited
      description: Decision Chamber Outcome Preview
...